requirements.txt
future>=0.18.2
Pillow>=9.1.1
screeninfo~=0.8.1
pyinstaller~=5.13.0
python-docx~=0.8.11
requests>=2.20.0